move state provider extensions to tools folder
