+++ libs/common/src/tools/state/data-packer.abstraction.ts
